+++ src/Epiklesis/Data.hs
@@ -5,12 +5,12 @@ import qualified Foreign.Lua as Lua
 
 -- a generic set of text, box bool will
 -- draw a box behind it, pos in game coords
-data WinText = WinText { winPos ∷ (Float, Float)
+data WinText = WinText { winPos ∷ (Double,Double)
                        , winBox ∷ Bool
                        , winText ∷ String } deriving (Show, Eq)
 
 -- a generic tile, use to display a single sprite
-data WinTile = WinTile { winTilePos ∷ (Float, Float)
+data WinTile = WinTile { winTilePos ∷ (Double,Double)
                        , winTileTex ∷ String } deriving (Show, Eq)
 
 -- data struct of what can be on a window,
