+++ src/Paracletus/Data.hs
@@ -17,8 +17,8 @@ data ParacResult = ParacSuccess | ParacError | GLFWSuccess | GLFWError | VulkanS
 -- on the screen, ind is the texture atlas
 -- index, t is the texture index in 2 dimensions
 -- size is the size of the atlas
-data GTile = GTile { tPos   ∷ (Float, Float)
-                   , tScale ∷ (Float, Float)
+data GTile = GTile { tPos   ∷ (Double,Double)
+                   , tScale ∷ (Double,Double)
                    , tInd   ∷ (Int, Int)
                    , tSize  ∷ (Int, Int)
                    , tT     ∷ Int }
